requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-genai>=0.7.0
google-generativeai>=0.8.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.3.0
langgraph>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
PyPDF2>=3.0.0
odfpy>=1.4.0
psycopg2-binary>=2.9.7
